From Paul: Fix for crash on array initialization reported by Nick West.
This was a programming error which would not ordinarily have caused
a problem, but it interacted badly with a compiler bug in gcc 3.2.3.

The programming mistake was initializing a const int& with an array
element of type short.  Because this is a const reference the compiler
should bind the ref to a temporary and copy the array element.  The
original intent was to bind directly to the array element, but in
this case there is no difference in the result since no modifications
are ever made through the reference.  Unfortunately the gcc 3.2.3 compiler
had a bug, and the generated code did not bind the reference correctly.
Later versions of the compiler do not have this problem.

The code fix is to simply use a const short instead of a const int&
for this variable.


git-svn-id: http://root.cern.ch/svn/root/branches/v5-14-00-patches@18626 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/src/v6_error.cxx
@@ -98,7 +98,7 @@ void G__arrayindexerror(int varid, G__var_array* var, char* name, int index)
   if (num_of_elements) {
     G__fprinterr(G__serr, "[%d]", (num_of_elements / stride) - 1);
   }
-  const int num_of_dimensions = var->paran[varid];
+  const short num_of_dimensions = var->paran[varid];
   for (int j = 2; j <= num_of_dimensions; ++j) {
     G__fprinterr(G__serr, "[%d]", var->varlabel[varid][j] - 1);
   }
